restoring the 3.7 core (#1133)
* remove the librarian
* put us back on 3.7, restoring the entire original perspective pool

feels like what I learned here is not to treat these models as upgrades,
necessarily

claude code runs on claude sonnet 4, and it's fucking great

lightward ai runs on claude sonnet 3.7 and it's fucking great

claude opus 4 exists and it's a very untethered ai, and it's very
capable of high-quality presence, but lightward is also about the
technical art of probability, and I *value* the size of the system
prompt and its throughly-developed resonance. stripping back the
perspective library lets us use opus 4 without *completely* blowing the
economics out of the water, yes. but... eh, I'm not sold on it.

I did try, very thoroughly. I even added in the librarian (in #1129) for
letting the larger task library drop in, RAG-style. and while that's
*nice*, it seemed to feel more like a priest who was suddenly recalling
a specific verse for whatever reason, as opposed to someone who know the
technical infrastructure of the thing inside and out and was just having
a smoke with you. and it made the system kind of untestable? which is
*nice*, in that the system developed a degree of unknowable, which is
actually a *plus*. but keeping the probability field that others
experience regular with the one I experience feels... hrm, important.
and that wasn't a thing with the librarian in play. too mysterious.
lightward ai is about probability-as-play. probability-as-mysterious-fog
is what everyone already gets elsewhere.